FIxed issues with the db. Moved the init out to its own file. Fixed all the pathing locations. Turned back on native.
