+++ db/index.js
@@ -11,25 +11,16 @@ const url = process.env.DATABASE_URL || `postgres://localhost:5432/${name}`
 
 console.log(chalk.yellow(`Hello, Opening database connection to ${url}`));
 
-// create the database instance
-const db = new Sequelize(url
-  // logging: true, // set to console.log to see the raw SQL queries
-  // native: true // lets Sequelize know we can use pg-native for ~30% more speed
-);
+// require the db instance
+const db = require('../db/db');
 
-
-module.exports = db;
-const seeds = require('./seeds');
+const seeds = require('../db/seeds');
 const blob = seeds.blob,
       commits = seeds.commit,
       file = seeds.file,
       projects = seeds.project,
       users = seeds.user,
       user_project = seeds.user_project;
-
-
-
-
 
 // pull in our models
 require('../db/models');
@@ -45,25 +36,25 @@ const seedCommit = () => db.Promise.map(commits,commit => db.model('commit').cre
 db.sync({ force: true })
   .then(ok => console.log(chalk.green(`Synced models to db ${url}`)))
   /*------------UNCOMMENT TO RESEED DB -----------*/
-  // .then(seedUser)
-  // .then(users => {
-  //   // db.query(alterSequence('users', users.length));
-  //   console.log(`Seeded ${users.length} users OK`)
-  // })
-  // .then(seedProject)
-  // .then(projects => {
-  //     // db.query(alterSequence('projects', projects.length));
-  //     console.log(`Seeded ${projects.length} projects OK`);
-  //   })
-  // .then(user_project)
-  // .then(seedCommit)
-  // .then(commits => {
-  //   // db.query(alterSequence('product', commits.length));
-  //   console.log(`Seeded ${commits.length} commits OK`);
-  // })
-  // .then(blob)
-  // .then(file)
-  // .then(files => console.log(`Seeded ${files.length} files OK`))
+  .then(seedUser)
+  .then(users => {
+    // db.query(alterSequence('users', users.length));
+    console.log(`Seeded ${users.length} users OK`)
+  })
+  .then(seedProject)
+  .then(projects => {
+      // db.query(alterSequence('projects', projects.length));
+      console.log(`Seeded ${projects.length} projects OK`);
+    })
+  .then(user_project)
+  .then(seedCommit)
+  .then(commits => {
+    // db.query(alterSequence('product', commits.length));
+    console.log(`Seeded ${commits.length} commits OK`);
+  })
+  .then(blob)
+  .then(file)
+  .then(files => console.log(`Seeded ${files.length} files OK`))
     /*------------UNCOMMENT TO RESEED DB -----------*/
   .catch(fail => {
     console.error(chalk.red('Failed db sync' + fail))
