+++ db/models/blob.js
@@ -1,7 +1,7 @@
 'use strict';
 console.log("BLOG FILE DIRNAME " + __dirname)
 const Sequelize = require('sequelize');
-const db = require('../index.js');
+const db = require('../db');
 console.log(db)
 const Blob = db.define('blob', {
   author: {
