+++ db/models/commit.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Sequelize = require('sequelize');
-const db = require('../index.js');
+const db = require('../db');
 
 const Commit = db.define('commit', {
   date: Sequelize.DATE,
