+++ db/models/project.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Sequelize = require('sequelize');
-const db = require('../index.js');
+const db = require('../db');
 
 const Project = db.define('project', {
   name: Sequelize.STRING
