+++ db/models/user.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Sequelize = require('sequelize');
-const db = require('../index.js');
+const db = require('../db');
 
 const User = db.define('user', {
   first_name: Sequelize.STRING,
