+++ db/models/user_project.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Sequelize = require('sequelize');
-const db = require('../index.js');
+const db = require('../db');
 
 const UserProject = db.define('user_project', {
   role: {
