+++ package.json
@@ -157,6 +157,7 @@
     "chalk": "^1.1.3",
     "chance": "^1.0.4",
     "electron-debug": "^1.0.1",
+    "electron-rebuild": "^1.4.0",
     "font-awesome": "^4.7.0",
     "morgan": "^1.7.0",
     "pg": "^6.1.0",
